Split root casper tests to multiple folders
