+++ casper/src/test/scala/coop/rchain/casper/api/BondedStatusAPITest.scala
@@ -9,9 +9,9 @@ import coop.rchain.casper.helper.BlockGenerator._
 import coop.rchain.casper.helper.TestNode.Effect
 import coop.rchain.casper.helper._
 import coop.rchain.casper.util.GenesisBuilder._
-import coop.rchain.casper.EngineWithCasper
 import coop.rchain.casper.util.ConstructDeploy
 import coop.rchain.casper.util.ConstructDeploy.basicDeployData
+import coop.rchain.casper.batch2.EngineWithCasper
 import coop.rchain.crypto.PublicKey
 import coop.rchain.crypto.signatures.Secp256k1
 import coop.rchain.metrics.Metrics
