+++ casper/src/test/scala/coop/rchain/casper/api/ExploratoryDeployAPITest.scala
@@ -7,7 +7,7 @@ import coop.rchain.casper.helper.{BlockDagStorageFixture, BlockGenerator, TestNo
 import coop.rchain.casper.util.GenesisBuilder.{buildGenesis, buildGenesisParameters}
 import coop.rchain.shared.scalatestcontrib.effectTest
 import coop.rchain.casper.engine.Engine
-import coop.rchain.casper.{EngineWithCasper, SafetyOracle}
+import coop.rchain.casper.SafetyOracle
 import coop.rchain.casper.helper.BlockGenerator._
 import coop.rchain.casper.util.ConstructDeploy.{basicDeployData, sourceDeployNowF}
 import coop.rchain.metrics.Metrics
@@ -15,6 +15,7 @@ import coop.rchain.shared.{Cell, Log}
 import coop.rchain.models._
 import coop.rchain.models.Expr.ExprInstance.GString
 import coop.rchain.casper.PrettyPrinter
+import coop.rchain.casper.batch2.EngineWithCasper
 import monix.eval.Task
 import org.scalatest.{EitherValues, FlatSpec, Matchers}
 import monix.execution.Scheduler.Implicits.global
