+++ casper/src/test/scala/coop/rchain/casper/batch1/MultiParentCasperReportingSpec.scala
@@ -1,5 +1,6 @@
-package coop.rchain.casper
+package coop.rchain.casper.batch1
 
+import coop.rchain.casper.ReportingCasper
 import coop.rchain.casper.ReportingCasperData.BlockTracesReport
 import coop.rchain.casper.helper.TestNode
 import coop.rchain.casper.helper.TestNode.Effect
@@ -7,8 +8,8 @@ import coop.rchain.casper.util.ConstructDeploy
 import coop.rchain.models.BlockHash._
 import coop.rchain.p2p.EffectsTestInstances.LogicalTime
 import coop.rchain.shared.scalatestcontrib.effectTest
-import org.scalatest.{FlatSpec, Inspectors, Matchers}
 import monix.execution.Scheduler.Implicits.global
+import org.scalatest.{FlatSpec, Inspectors, Matchers}
 
 class MultiParentCasperReportingSpec extends FlatSpec with Matchers with Inspectors {
