+++ casper/src/test/scala/coop/rchain/casper/batch1/MultiParentCasperRholangSpec.scala
@@ -1,13 +1,14 @@
-package coop.rchain.casper
+package coop.rchain.casper.batch1
 
 import coop.rchain.casper.helper.TestNode
 import coop.rchain.casper.helper.TestNode.Effect
-import coop.rchain.shared.scalatestcontrib._
 import coop.rchain.casper.util.rholang.{RegistrySigGen, RuntimeManager}
 import coop.rchain.casper.util.{ConstructDeploy, ProtoUtil, RSpaceUtil}
+import coop.rchain.casper.{Created, MultiParentCasper, MultiParentCasperImpl}
 import coop.rchain.crypto.codec.Base16
 import coop.rchain.crypto.signatures.Secp256k1
 import coop.rchain.p2p.EffectsTestInstances.LogicalTime
+import coop.rchain.shared.scalatestcontrib._
 import monix.execution.Scheduler.Implicits.global
 import org.scalatest.{FlatSpec, Inspectors, Matchers}
