+++ casper/src/test/scala/coop/rchain/casper/batch1/MultiParentCasperSmokeSpec.scala
@@ -1,11 +1,11 @@
-package coop.rchain.casper
+package coop.rchain.casper.batch1
 
 import cats.implicits._
 import coop.rchain.casper.helper.TestNode
 import coop.rchain.casper.helper.TestNode._
-import coop.rchain.shared.scalatestcontrib._
 import coop.rchain.casper.util.ConstructDeploy
 import coop.rchain.p2p.EffectsTestInstances.LogicalTime
+import coop.rchain.shared.scalatestcontrib._
 import monix.execution.Scheduler.Implicits.global
 import org.scalatest.{FlatSpec, Inspectors, Matchers}
