+++ casper/src/test/scala/coop/rchain/casper/batch2/CliqueOracleTest.scala
@@ -1,19 +1,21 @@
-package coop.rchain.casper
+package coop.rchain.casper.batch2
 
-import scala.collection.immutable.HashMap
-import scala.collection.mutable
+import com.google.protobuf.ByteString
 import coop.rchain.blockstorage.BlockStore
-import coop.rchain.metrics.{Metrics, NoopSpan, Span}
+import coop.rchain.blockstorage.dag.IndexedBlockDagStorage
+import coop.rchain.casper.SafetyOracle
 import coop.rchain.casper.helper.BlockGenerator._
 import coop.rchain.casper.helper.BlockUtil.generateValidator
 import coop.rchain.casper.helper.{BlockDagStorageFixture, BlockGenerator}
 import coop.rchain.casper.protocol.{BlockMessage, Bond}
+import coop.rchain.metrics.{Metrics, NoopSpan, Span}
 import coop.rchain.models.Validator.Validator
 import coop.rchain.p2p.EffectsTestInstances.LogStub
-import com.google.protobuf.ByteString
-import coop.rchain.blockstorage.dag.IndexedBlockDagStorage
 import monix.eval.Task
 import org.scalatest.{FlatSpec, Matchers}
+
+import scala.collection.immutable.HashMap
+import scala.collection.mutable
 
 class CliqueOracleTest
     extends FlatSpec
