+++ casper/src/test/scala/coop/rchain/casper/batch2/SingleParentCasperSpec.scala
@@ -1,19 +1,20 @@
-package coop.rchain.casper
+package coop.rchain.casper.batch2
 
-import cats.syntax.traverse._
-import cats.syntax.flatMap._
-import cats.syntax.either._
 import cats.instances.list._
+import cats.syntax.either._
+import cats.syntax.flatMap._
+import cats.syntax.traverse._
 import coop.rchain.casper.helper.TestNode
 import coop.rchain.casper.helper.TestNode.Effect
-import coop.rchain.casper.protocol.{DeployData, Justification}
+import coop.rchain.casper.protocol.DeployData
 import coop.rchain.casper.util.ConstructDeploy
 import coop.rchain.casper.util.GenesisBuilder.buildGenesis
+import coop.rchain.casper.{BlockStatus, ValidBlock, Validate}
 import coop.rchain.crypto.signatures.Signed
 import coop.rchain.p2p.EffectsTestInstances.LogicalTime
 import coop.rchain.shared.scalatestcontrib.effectTest
-import org.scalatest.{FlatSpec, Inspectors, Matchers}
 import monix.execution.Scheduler.Implicits.global
+import org.scalatest.{FlatSpec, Inspectors, Matchers}
 
 class SingleParentCasperSpec extends FlatSpec with Matchers with Inspectors {
   implicit val timeEff = new LogicalTime[Effect]
